Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@4814 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ GDiffView/src/nu/mine/mosher/GDiffView.java
@@ -390,7 +390,7 @@ public class GDiffView extends JFrame
         return getAscStart(pos) + 1;
     }
 
-    public void highlight(int beginPoint, int endPoint, boolean highlight)
+    public void highlight(long beginPoint, long endPoint, boolean highlight)
     {
         AttributeSet attr = (AttributeSet)styles.get(highlight
                 ? "highlight"
